import/export works

+++ app/src/main/java/hu/bme/aut/android/costofliving/file/FileHandler.kt
@@ -1,10 +1,7 @@
 package hu.bme.aut.android.costofliving.file
 
 import android.content.Context
-import android.content.Intent
 import android.os.Environment
-import android.util.Log
-import de.raphaelebner.roomdatabasebackup.core.RoomBackup
 import hu.bme.aut.android.costofliving.data.ExpenseItem
 import hu.bme.aut.android.costofliving.data.ExpenseListDatabase
 import hu.bme.aut.android.costofliving.data.LoanItem
@@ -16,8 +13,11 @@ import kotlin.concurrent.thread
 class FileHandler(context: Context, username: String) {
     private val applicationContext: Context = context
     private val user: String = username
+    private val expenseDBFieldNames = "id,name,description,category,cost,is_expense,username,year,month,is_shared"
+    private val loansDBFieldNames = "id,loaner_name,amount,description,username,is_loaner"
 
-    fun exportToCSV() {
+    //for testing export/import
+    /*fun addTestItem(){
         thread {
             val state = Environment.getExternalStorageState()
             if (!state.equals(Environment.MEDIA_MOUNTED)) {
@@ -28,18 +28,66 @@ class FileHandler(context: Context, username: String) {
             if (!myDir.exists()) {
                 myDir.mkdirs()
             }
+            val filename = "expenses_${user}.csv"
+            val file = File(myDir, filename)
+            var test = "nice, and good"
+            test = escapeComma(test)
+            try {
+                val out = FileOutputStream(file, true).bufferedWriter().use { out->
+                    out.append("111,escaped item,${test},Food,111,TRUE,test,2022,4,FALSE\n")
+                }
+                out.close()
+            } catch (e: Exception) {
+                e.printStackTrace()
+            }
+        }
+    }*/
 
-            //export EXPENSES database
-            var filename = "expenses_${user}.csv"
-            var file = File(myDir, filename)
+    private fun escapeComma(rawString: String): String{
+        // this is not elegant, but works
+        // "d11c0182" is the CRC32 hash of "comma"
+        var raw = rawString
+        if (raw.contains(",")) {
+            raw = raw.replace(",", "d11c0182")
+        }
+        return raw
+    }
+
+    private fun recoverComma(rawString: String): String{
+        // this is not elegant, but works
+        // "d11c0182" is the CRC32 hash of "comma"
+        var raw = rawString
+        if (raw.contains("d11c0182")) {
+            raw = raw.replace("d11c0182", ",")
+        }
+        return raw
+    }
+
+    fun exportExpensesDBToCSV() {
+        thread {
+            val state = Environment.getExternalStorageState()
+            if (!state.equals(Environment.MEDIA_MOUNTED)) {
+                return@thread
+            }
+            val root = applicationContext.getExternalFilesDir(null)
+            val myDir = File("$root/exported_databases")
+            if (!myDir.exists()) {
+                myDir.mkdirs()
+            }
+            val filename = "expenses_${user}.csv"
+            val file = File(myDir, filename)
             if (file.exists()) file.delete()
             val expenseDB = ExpenseListDatabase.getDatabase(applicationContext)
             val expenseItems = expenseDB.expenseItemDao().getAll(user)
             try {
                 val out = FileOutputStream(file)
                 file.printWriter().use { out ->
-                    out.println("id,name,description,category,cost,is_expense,username,year,month,is_shared")
+                    out.println(expenseDBFieldNames)
                     expenseItems.forEach {
+                        it.name = escapeComma(it.name)
+                        it.description = escapeComma(it.description)
+                        it.category = escapeComma(it.category)
+                        it.username = escapeComma(it.username)
                         out.println(
                             "${it.id},${it.name},${it.description},${it.category},${it.cost}," +
                                     "${it.isExpense},${it.username},${it.year},${it.month},${it.isShared}"
@@ -51,18 +99,33 @@ class FileHandler(context: Context, username: String) {
             } catch (e: Exception) {
                 e.printStackTrace()
             }
+        }
+    }
 
-            //export LOANS database
-            filename = "loans_${user}.csv"
-            file = File(myDir, filename)
+    fun exportLoansDBToCSV() {
+        thread {
+            val state = Environment.getExternalStorageState()
+            if (!state.equals(Environment.MEDIA_MOUNTED)) {
+                return@thread
+            }
+            val root = applicationContext.getExternalFilesDir(null)
+            val myDir = File("$root/exported_databases")
+            if (!myDir.exists()) {
+                myDir.mkdirs()
+            }
+            val filename = "loans_${user}.csv"
+            val file = File(myDir, filename)
             if (file.exists()) file.delete()
             val loanDB = LoanListDatabase.getDatabase(applicationContext)
             val loanItems = loanDB.loanItemDao().getAll(user)
             try {
                 val out = FileOutputStream(file)
                 file.printWriter().use { out ->
-                    out.println("id,loaner_name,amount,description,username,is_loaner")
+                    out.println(loansDBFieldNames)
                     loanItems.forEach {
+                        it.loanerName = escapeComma(it.loanerName)
+                        it.description = escapeComma(it.description)
+                        it.username = escapeComma(it.username)
                         out.println(
                             "${it.id},${it.loanerName},${it.amount},${it.description}," +
                                     "${it.username},${it.isLoaner}"
@@ -77,52 +140,79 @@ class FileHandler(context: Context, username: String) {
         }
     }
 
-    /*fun restoreExpenseDB(fileName: String) {
-       thread {
-           val inputStream: InputStream = File(fileName).inputStream()
-           val lineList = mutableListOf<String>()
-           inputStream.bufferedReader().forEachLine { lineList.add(it) }
-           val expenseItems = mutableListOf<ExpenseItem>()
-           lineList.forEach {
-               val arguments = it.split(',')
-               val expenseItem = ExpenseItem(
-                   arguments[0].toLong(), arguments[1], arguments[2],
-                   arguments[3], arguments[4].toInt(), arguments[5].toBoolean(), arguments[6],
-                   arguments[7].toInt(), arguments[8].toInt(), arguments[9].toBoolean()
-               )
-               expenseItems.add(expenseItem)
-           }
-           if (expenseItems.size > 0) {
-               val database = ExpenseListDatabase.getDatabase(applicationContext)
-               database.expenseItemDao().clearTable()
-               for (item in expenseItems)
-                   database.expenseItemDao().insert(item)
-           }
-       }
-   }*/
-
-    /*fun restoreLoanDB(fileName: String) {
+    fun importExpensesDBFromCSV() {
         thread {
-            val inputStream: InputStream = File(fileName).inputStream()
-            val lineList = mutableListOf<String>()
-            inputStream.bufferedReader().forEachLine { lineList.add(it) }
-            val loanItems = mutableListOf<LoanItem>()
-            lineList.forEach {
-                val arguments = it.split(',')
-                val loanItem = LoanItem(
-                    arguments[0].toLong(), arguments[0], arguments[0].toInt(), arguments[0],
-                    arguments[0], arguments[0].toBoolean()
-                )
-                loanItems.add(loanItem)
+            val state = Environment.getExternalStorageState()
+            if (!state.equals(Environment.MEDIA_MOUNTED)) {
+                return@thread
             }
-            if (loanItems.size > 0) {
-                val database = LoanListDatabase.getDatabase(applicationContext)
-                database.loanItemDao().clearTable()
-                for (item in loanItems)
-                    database.loanItemDao().insert(item)
+            val root = applicationContext.getExternalFilesDir(null)
+            val myDir = File("$root/exported_databases")
+            val filename = "expenses_${user}.csv"
+            val file = File(myDir, filename)
+            if (!file.exists()) return@thread
+            val expenseDB = ExpenseListDatabase.getDatabase(applicationContext)
+            expenseDB.expenseItemDao().clearTable(user)
+            try {
+                val fis = FileInputStream(file)
+                file.forEachLine {
+                    if(it != expenseDBFieldNames){
+                        val subStrings = it.split(',')
+                        val expenseItem = ExpenseItem(
+                            name = recoverComma(subStrings[1]),
+                            description = recoverComma(subStrings[2]),
+                            cost = subStrings[4].toInt(),
+                            category =  recoverComma(subStrings[3]),
+                            isExpense = subStrings[5].toBoolean(),
+                            username = recoverComma(subStrings[6]),
+                            year = subStrings[7].toInt(),
+                            month = subStrings[8].toInt(),
+                            isShared = subStrings[9].toBoolean()
+                        )
+                        expenseDB.expenseItemDao().insert(expenseItem)
+                    }
+                }
+                fis.close()
+            } catch (e: Exception) {
+                e.printStackTrace()
             }
         }
-    }*/
+    }
+
+    fun importLoansDBFromCSV() {
+        thread {
+            val state = Environment.getExternalStorageState()
+            if (!state.equals(Environment.MEDIA_MOUNTED)) {
+                return@thread
+            }
+            val root = applicationContext.getExternalFilesDir(null)
+            val myDir = File("$root/exported_databases")
+            val filename = "loans_${user}.csv"
+            val file = File(myDir, filename)
+            if (!file.exists()) return@thread
+            val loanDB = LoanListDatabase.getDatabase(applicationContext)
+            loanDB.loanItemDao().clearTable(user)
+            try {
+                val fis = FileInputStream(file)
+                file.forEachLine {
+                    if(it != loansDBFieldNames){
+                        val subStrings = it.split(',')
+                        val loanItem = LoanItem(
+                            loanerName = recoverComma(subStrings[1]),
+                            description = recoverComma(subStrings[3]),
+                            amount = subStrings[2].toInt(),
+                            isLoaner = subStrings[5].toBoolean(),
+                            username = recoverComma(subStrings[4])
+                        )
+                        loanDB.loanItemDao().insert(loanItem)
+                    }
+                }
+                fis.close()
+            } catch (e: Exception) {
+                e.printStackTrace()
+            }
+        }
+    }
 
     /*fun backupDB() {
         thread {
